Move commander error probe from shared utils to CLI module
Relocated isCommanderErrorLike from src/shared/utils/capability-probes.js into src/cli/lib/commander-error-utils.js so the Commander-specific detection stays within the CLI layer; updated imports and tests.
